restore: Add option warm-up-wait
